requirements.txt
google-genai
python-dotenv
tenacity
genanki
pydantic
graphviz

python-dotenv
tenacity
